Fix issue with multiple matching gem versions.
A gem sometimes have multiple variations of the same version intended
for different platforms. Before these changes, resona couldn't handle
this case very well and returned the wrong checksum in some cases. But
as it turned out, we were able to ask Bundler for the correct checksum
of the gem we wanted.

To sum up, these changes

  - Removes manual queries to rubygems.org for checksums
  - Asks Bundler for the checksums instead